Create a separate class for serialization
